rename DynamicChunkRecordWriter to VectorStoreWriter to expose only the most powerful utility to the end users
